+++ Scrbl.JaccoBikker/Utility/Ray.cs
@@ -4,23 +4,23 @@
 
     public struct Ray
     {
-        public Vector3d Origin { get { return O; } set { O = value; } }
-        public Vector3d Direction { get { return D; } set { D = value; } }
-        public double Time { get { return T; } set { T = value; } }
-        public double t { get { return T; } set { T = value; } }
-        public Vector3d O;
-        public Vector3d D;
-        public double T;
+        public Vector3f Origin { get { return O; } set { O = value; } }
+        public Vector3f Direction { get { return D; } set { D = value; } }
+        public float Time { get { return T; } set { T = value; } }
+        public float t { get { return T; } set { T = value; } }
+        public Vector3f O;
+        public Vector3f D;
+        public float T;
 
         // Parameterless constructor (Defaults to 0,0,0 vectors)
         public Ray()
         {
-            Origin = new Vector3d();
-            Direction = new Vector3d();
+            Origin = new Vector3f();
+            Direction = new Vector3f();
         }
 
         // Main constructor
-        public Ray(Vector3d origin, Vector3d direction, double time = 0.0)
+        public Ray(Vector3f origin, Vector3f direction, float time = 0.0f)
         {
             Origin = origin;
             Direction = direction;
@@ -28,7 +28,7 @@
         }
 
         // Linear interpolation along the ray: orig + t * dir
-        public readonly Vector3d At(double t)
+        public readonly Vector3f At(float t)
         {
             return Origin + Direction * t;
         }
