+++ Scrbl.JaccoBikker/Utility/Triangle.cs
@@ -4,12 +4,12 @@
 
     public class Triangle
     {
-        public Vector3d A { get { return vertex0; } set { vertex0 = value; } }
-        public Vector3d B { get { return vertex1; } set { vertex1 = value; } }
-        public Vector3d C { get { return vertex2; } set { vertex2 = value; } }
+        public Vector3f A { get { return vertex0; } set { vertex0 = value; } }
+        public Vector3f B { get { return vertex1; } set { vertex1 = value; } }
+        public Vector3f C { get { return vertex2; } set { vertex2 = value; } }
 
-        public Vector3d vertex0 = new Vector3d();
-        public Vector3d vertex1 = new Vector3d();
-        public Vector3d vertex2 = new Vector3d();
+        public Vector3f vertex0 = new Vector3f();
+        public Vector3f vertex1 = new Vector3f();
+        public Vector3f vertex2 = new Vector3f();
     }
 }
